Reorganize formatters into separate files
