+++ src/Setlistbot.Domain.GratefulDead/SetlistsFormatter.cs
@@ -1,0 +1,19 @@
+using Setlistbot.Domain.Formatters;
+
+namespace Setlistbot.Domain.GratefulDead
+{
+    public sealed class SetlistsFormatter(IEnumerable<Setlist> setlists) : IFormatter
+    {
+        public string Format() =>
+            setlists
+                .Select(setlist => new CombinedFormatter(
+                    new ArchiveOrgLinkFormatter(setlist),
+                    new SpaceFormatter(),
+                    new CharacterFormatter('@'),
+                    new SpaceFormatter(),
+                    new LocationFormatter(setlist.Location),
+                    new NewLineFormatter(2)
+                ))
+                .Format();
+    }
+}
