+++ src/Setlistbot.Domain.Phish/PhishInLinkFormatter.cs
@@ -1,0 +1,10 @@
+﻿using Setlistbot.Domain.Formatters;
+
+namespace Setlistbot.Domain.Phish
+{
+    public sealed class PhishInLinkFormatter(DateOnly date, IFormatter textFormatter) : IFormatter
+    {
+        public string Format() =>
+            new MarkdownLinkFormatter(PhishUri.PhishIn(date), textFormatter).Format();
+    }
+}
