+++ src/Setlistbot.Domain/Formatters/CharacterFormatter.cs
@@ -1,0 +1,10 @@
+﻿namespace Setlistbot.Domain.Formatters
+{
+    /// <summary>
+    /// Formats a character as-is.
+    /// </summary>
+    public sealed class CharacterFormatter(char c) : IFormatter
+    {
+        public string Format() => c.ToString();
+    }
+}
