+++ src/Setlistbot.Domain/Formatters/EmptyFormatter.cs
@@ -1,0 +1,10 @@
+﻿namespace Setlistbot.Domain.Formatters
+{
+    /// <summary>
+    /// Returns an empty string.
+    /// </summary>
+    public sealed class EmptyFormatter : IFormatter
+    {
+        public string Format() => string.Empty;
+    }
+}
