+++ src/Setlistbot.Domain/Formatters/SetFormatter.cs
@@ -14,15 +14,4 @@
                 new SongsFormatter(set.Songs)
             ).Format();
     }
-
-    public sealed class SetsFormatter(IEnumerable<Set> sets) : IFormatter
-    {
-        public string Format() =>
-            sets.Select<Set, IFormatter>(s => new CombinedFormatter(
-                    new SetFormatter(s),
-                    new NewLineFormatter(2)
-                ))
-                .ToArray()
-                .Format();
-    }
 }
