+++ src/Setlistbot.Domain/Formatters/SongTransitionFormatter.cs
@@ -11,23 +11,4 @@
                 _ => throw new NotImplementedException(),
             };
     }
-
-    public sealed class SongTransitionSuffixFormatter(SongTransition songTransition) : IFormatter
-    {
-        public string Format() =>
-            (
-                songTransition switch
-                {
-                    SongTransition.Stop => new CombinedFormatter(
-                        new SongTransitionFormatter(songTransition),
-                        new SpaceFormatter()
-                    ),
-                    _ => new CombinedFormatter(
-                        new SpaceFormatter(),
-                        new SongTransitionFormatter(songTransition),
-                        new SpaceFormatter()
-                    ),
-                }
-            ).Format();
-    }
 }
